third_party/WebKit/LayoutTests/external/wpt/html/webappapis/system-state-and-capabilities/the-navigator-object/get-navigatorlanguage-manual-expected.txt

This is a testharness.js-based test.
FAIL NavigatorLanguage: navigator.language returns the user's preferred languageassert_equals: expected "en" but got "en-US"
Harness: the test ran to completion.
